Tried to test, updated rotateLeft extension
I tried, really really tried, to get proper testing setup but there was some weird problem where the package was not recognized for testing.  I thought it might have been because I renamed things but tried the same procedure on the default "Hello" executable package but the same problem cropped up.  Maybe I could try it with just a library package.... but i dont care that much,,, or have that kind of time.

ALSO, I was alerted to the fact that rotating a 16-bit number is not the same as rotating a 5-bit number inside a 16-bit number so I fixed (hopefully) that.

Testing of all of this will be in a poor-boi's testing class that I will create and run via the main module/executable.

Good luck future me, and remember to eat, sleep, take short breaks, work hard and play after it's done.